Implementaciones y mejoras generales:
- Modificado el formulario de turno para que al seleccionar un servicio
se carguen los empleados del establecimiento correspondiente.
- Ajustado el flujo para que un cliente, al pedir turno, no pueda
seleccionar otro cliente sino que se utilice automáticamente el de la
sesión.
- Añadidas excepciones personalizadas para manejar errores 403, 500 y
errores genéricos.
- Implementada validación para evitar la eliminación de establecimientos
con servicios asociados y otras validaciones similares.
- Refuerzo en la seguridad de rutas: asegurado que sólo usuarios con
privilegios adecuados puedan acceder a ciertas URL.
- Correcciones y mejoras menores en la lógica y presentación.

+++ src/main/java/com/oo2/grupo20/services/implementation/EstablecimientoService.java
@@ -17,6 +17,7 @@ import com.oo2.grupo20.services.IEstablecimientoService;
 import com.oo2.grupo20.repositories.IEmpleadoRepository;
 import com.oo2.grupo20.repositories.IEstablecimientoRepository;
 import com.oo2.grupo20.exceptions.EstablecimientoConEmpleadosException;
+import com.oo2.grupo20.exceptions.EstablecimientoConServiciosException;
 
 @Service ("establecimientoService")
 public class EstablecimientoService implements IEstablecimientoService{
@@ -62,15 +63,20 @@ private IEstablecimientoRepository establecimientoRepository;
 	        Establecimiento est = optionalEstablecimiento.get();
 
 	        if (!est.getEmpleados().isEmpty()) {
-	            throw new EstablecimientoConEmpleadosException("No se puede eliminar el establecimiento con empleados asociados.");
+	            throw new EstablecimientoConEmpleadosException("No se puede eliminar un establecimiento con empleados asociados.");
+	        }
+
+	        if (est.getServicios() != null && !est.getServicios().isEmpty()) {
+	            throw new EstablecimientoConServiciosException("No se puede eliminar un establecimiento con servicios asociados.");
 	        }
 
 	        establecimientoRepository.deleteById(idEstablecimiento);
 	        return true;
 	    }
 
-	    return false; // no se encontró
+	    return false;
 	}
+
 	
 	
 	@Override
